api CRUD post

+++ app/Http/Controllers/Api/PostController.php
@@ -7,6 +7,8 @@ use Illuminate\Http\Request;
 use App\Models\Post;
 use App\Models\File;
 use App\Models\User;
+use App\Models\Like;
+use Illuminate\Support\Facades\Storage;
 
 class PostController extends Controller
 {
@@ -15,19 +17,21 @@ class PostController extends Controller
      */
     public function index()
     {
-        //
-        $posts = Post::all();
+        if ( $this->authorize('viewAny', Post::class)){
+            
+            $posts = Post::all();
 
-        if ($posts) {
-            return response()->json([
-                'success' => true,
-                'data' => $posts
-            ], 200);
-        } else {
-            return response()->json([
-                'success'  => false,
-                'message' => 'Error list posts'
-            ], 500);
+            if ($posts) {
+                return response()->json([
+                    'success' => true,
+                    'data' => $posts
+                ], 200);
+            } else {
+                return response()->json([
+                    'success'  => false,
+                    'message' => 'Error list posts'
+                ], 500);
+            }
         }
     }
 
@@ -35,57 +39,58 @@ class PostController extends Controller
      * Store a newly created resource in storage.
      */
     public function store(Request $request)
-    {
-        //
-        $validatedData = $request->validate([
-            'title' => 'required|max:20',
-            'description' => 'required|max:200',
-            'upload' => 'required|mimes:gif,jpeg,jpg,png|max:1024',
-            'visibility_id' => 'required|exists:visibilities,id', 
-        ]);
-        
-        // Obtención de los datos del formulario.
-        $upload = $request->file('upload');
-        $fileName = $upload->getClientOriginalName();
-        $fileSize = $upload->getSize();
-
-        // Almacenamiento del archivo en el sistema de archivos y la base de datos.
-        $uploadName = time() . '_' . $fileName;
-        $filePath = $upload->storeAs(
-            'uploads',
-            $uploadName,
-            'public'
-        );
-
-        if (\Storage::disk('public')->exists($filePath)) {
-
-            $fullPath = \Storage::disk('public')->path($filePath);
-
-            // Creación de la entrada del archivo en la base de datos.
-            $file = File::create([
-                'filepath' => $filePath,
-                'filesize' => $fileSize,
-            ]);
-
-            // Creación de la publicación en la base de datos.
-            $post = Post::create([
-                'author_id' => auth()->user()->id,
-                'file_id' => $file->id,
-                'title' => $request->title,
-                'description' => $request->description,
-                'visibility_id' => $request->visibility_id
+    {        
+        if ( $this->authorize('create', Post::class)) {
+            $validatedData = $request->validate([
+                'title' => 'required|max:20',
+                'description' => 'required|max:200',
+                'upload' => 'required|mimes:gif,jpeg,jpg,png|max:1024',
+                'visibility_id' => 'required|exists:visibilities,id', 
             ]);
             
-            return response()->json([
-                'success' => true,
-                'data'    => $post
-            ], 201);
+            // Obtención de los datos del formulario.
+            $upload = $request->file('upload');
+            $fileName = $upload->getClientOriginalName();
+            $fileSize = $upload->getSize();
 
-        } else {
-            return response()->json([
-                'success'  => false,
-                'message' => 'Error uploading post'
-            ], 500);
+            // Almacenamiento del archivo en el sistema de archivos y la base de datos.
+            $uploadName = time() . '_' . $fileName;
+            $filePath = $upload->storeAs(
+                'uploads',
+                $uploadName,
+                'public'
+            );
+
+            if (Storage::disk('public')->exists($filePath)) {
+
+                $fullPath = Storage::disk('public')->path($filePath);
+
+                // Creación de la entrada del archivo en la base de datos.
+                $file = File::create([
+                    'filepath' => $filePath,
+                    'filesize' => $fileSize,
+                ]);
+
+                // Creación de la publicación en la base de datos.
+                $post = Post::create([
+                    'author_id' => auth()->user()->id,
+                    'file_id' => $file->id,
+                    'title' => $request->title,
+                    'description' => $request->description,
+                    'visibility_id' => $request->visibility_id
+                ]);
+                
+                return response()->json([
+                    'success' => true,
+                    'data'    => $post
+                ], 201);
+
+            } else {
+                return response()->json([
+                    'success'  => false,
+                    'message' => 'Error uploading post'
+                ], 500);
+            }
         }
     }
 
@@ -95,12 +100,14 @@ class PostController extends Controller
     public function show(string $id)
     {
         $post = Post::find($id);
-
+        
         if ($post) {
-            return response()->json([
-                'success' => true,
-                'data' => $post
-            ], 200);
+            if ($this->authorize('view', $post)) {
+                return response()->json([
+                    'success' => true,
+                    'data' => $post
+                ], 200);
+            }
         } else {
             return response()->json([
                 'success'  => false,
@@ -114,48 +121,49 @@ class PostController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        //
         $post = Post::find($id);
         if ($post) {
-            $validatedData = $request->validate([
-                'title' => 'required|max:20',
-                'description' => 'required|max:200',
-                'upload' => 'required|mimes:gif,jpeg,jpg,png|max:1024',
-                'visibility_id' => 'required|exists:visibilities,id', 
-            ]);
-
-            if ($request->hasFile('upload')) {
-                \Storage::disk('public')->delete($post->file->filepath);
-    
-                $newFile = $request->file('upload');
-                $newFileName = time() . '_' . $newFile->getClientOriginalName();
-                $newFilePath = $newFile->storeAs('uploads', $newFileName, 'public');
-    
-                $post->file->update([
-                    'original_name' => $newFile->getClientOriginalName(),
-                    'filesize' => $newFile->getSize(),
-                    'filepath' => $newFilePath,
+            if ( $this->authorize('update', $post)) {
+                $validatedData = $request->validate([
+                    'title' => 'required|max:20',
+                    'description' => 'required|max:200',
+                    'upload' => 'required|mimes:gif,jpeg,jpg,png|max:1024',
+                    'visibility_id' => 'required|exists:visibilities,id', 
                 ]);
-            }
 
-            $post->update([
-                'author_id' => auth()->user()->id,
-                'file_id' => $post->file->id,
-                'title' => $request->title,
-                'description' => $request->description,
-                'visibility_id' => $request->visibility_id,
-            ]);
+                if ($request->hasFile('upload')) {
+                    Storage::disk('public')->delete($post->file->filepath);
+        
+                    $newFile = $request->file('upload');
+                    $newFileName = time() . '_' . $newFile->getClientOriginalName();
+                    $newFilePath = $newFile->storeAs('uploads', $newFileName, 'public');
+        
+                    $post->file->update([
+                        'original_name' => $newFile->getClientOriginalName(),
+                        'filesize' => $newFile->getSize(),
+                        'filepath' => $newFilePath,
+                    ]);
+                }
 
-            if ($post) {
-                return response()->json([
-                    'success' => true,
-                    'data'    => $post
-                ], 200);
-            } else {
-                return response()->json([
-                    'success'  => false,
-                    'message' => 'Error updating post'
-                ], 500);
+                $post->update([
+                    'author_id' => auth()->user()->id,
+                    'file_id' => $post->file->id,
+                    'title' => $request->title,
+                    'description' => $request->description,
+                    'visibility_id' => $request->visibility_id,
+                ]);
+
+                if ($post) {
+                    return response()->json([
+                        'success' => true,
+                        'data'    => $post
+                    ], 200);
+                } else {
+                    return response()->json([
+                        'success'  => false,
+                        'message' => 'Error updating post'
+                    ], 500);
+                }
             }
         } else {
             return response()->json([
@@ -169,16 +177,21 @@ class PostController extends Controller
      * Remove the specified resource from storage.
      */
     public function destroy(string $id)
-    {
-        //
+    {        
         $post = Post::find($id);
         if ($post) {
-            $post->delete();
+            if ( $this->authorize('delete', $post)) {
+                if (Storage::disk('public')->exists($post->file->filepath)) {
+                    Storage::disk('public')->delete($post->file->filepath);
+                }
+                $post->file->delete();
+                $post->delete();
 
-            return response()->json([
-                'success' => true,
-                'data' => $post
-            ], 200);
+                return response()->json([
+                    'success' => true,
+                    'data' => $post
+                ], 200);
+            }
         } else {
             return response()->json([
                 'success'  => false,
@@ -191,32 +204,53 @@ class PostController extends Controller
     {
         $post = Post::find($id);
         $user = auth()->user();
-        dd($user);
-        if ($post) {
+        if ($post){
+            if ( $this->authorize('like', $post)) {
+                if (!$user->likes->contains($post->id)) {
+                    $user->likes()->attach($post);
+                    return response()->json([
+                        'success' => true,
+                        'data'    => ['Post liked successfully']
+                    ], 200);
+                }
+            }
+    
             return response()->json([
-                'success' => true,
-                'data'    => $post
-            ], 200);
+                'success'  => false,
+                'message' => 'Post already liked'
+            ], 404);
         }
-
         return response()->json([
             'success'  => false,
-            'message' => 'Error like post'
-        ], 500);
+            'message' => 'Post not found'
+        ], 404);
     }
 
-    public function unlike(Post $post)
+    public function unlike(string $id)
     {
-        $this->authorize('create', $post);
+        $post = Post::find($id);
         $user = auth()->user();
 
-        // Verificar si el usuario ha dado "like" a esta publicación antes de intentar eliminarlo.
-        if ($user->likes->contains($post->id)) {
-            $user->likes()->detach($post);
-            return redirect()->route('posts.index')->with('success', 'Like eliminado');
+        if ($post){
+            if ($user->likes->contains($post->id)) {
+                $user->likes()->detach($post);
+                return response()->json([
+                    'success' => true,
+                    'data'    => ['Post unliked successfully']
+                ], 200);        
+            } else {
+                return response()->json([
+                    'success'  => false,
+                    'message' => 'Post not liked'
+                ], 404);
+            }
+                
+        } else {
+            return response()->json([
+                'success'  => false,
+                'message' => 'Post not found'
+            ], 404);
         }
-
-        return redirect()->route('posts.index')->with('error', 'No has dado like a esta publicación previamente');
     }
 
     public function update_workaround(Request $request, $id)
